Features: Started remote client, EEPROM manager rework
The old eeprom implementation was pretty dodgy, opted for a different (and hopefully more reliable) approach that will prevent wear on the eeprom. Passwords are now stored at the start in 96 byte blocks, and there is a 32 byte bitmask at the start which indicates the blocks that have been filled, so there will be space for 256 password entries. The wallets use the old system (minus the types), for more of a linked-list like system, as I doubt the average user will have lots of wallets/modify them much.

I also decided that manually editing the passwords with the IR remote is pretty slow and annoying, so I have started making a remote config functionality. It will use a session key as well as a master key to keep the data secure. The vault.py is a cli that allows you to add/delete/modify/read password entries, and eventually add/delete wallet entries. I will likely also implement a serial interface mode, so you can get the full cli functionality without connecting to a network.

The things I need to add at the moment is encryption in the remote comms on all the sensitive information (some things are encrypted, but not all). I also need to add the session key encryption to the remote comms. I also don't like that ALL of the credentials are decrypted on start up and sit on the heap, so I am going to modify the code to only decrypt the names on start up, and decrypt other info when it is needed. The 'authentication' for the remote functionality also needs replacing to something that actually provides protection (if you know the heap, then you can interact with it at the moment). Will probably use some sort of challenge from the esp to verify the client has the correct session and master keys. Also missing the wallet add functionality at the moment.