Fixes: Plaintext passwords generated only when needed, code modularisation
Made the code work so that the key is set in the aes instance when it is entered, and is not stored anywhere else. Passwords and wallet entries are no longer stored in plaintext, they are stored encrypted on the heap, and are only decrypted when they are selected to be viewed. 

Also tidied up the code slightly (made it more modular), removed some unused functionality, and gave the code a tiny clean up.