Fix: Adjusted max lengths, fixed identical encrypted blocks, added wallet delete
Added the final core remote command (wlt del) which deletes the wallet with the specified name.

Reduces the length of the wallet phrases so that they fit on the screen, the maximum size is now 20. Also, set the max saved password/username length to 24.

Changed the layout of the wallet display to be centre-aligned.

I realised the random padding at the end of the text to be encrypted would be useless if the password had a length > 16 as the first block would be unaffected by the random data appended to the end. Fixed this by splitting the text to be encrypted into 2 12-byte chunks, and appending 4 random bytes to the end of each chunk (on top of the random padding up to a length of 24), this significantly reduces the chance of collisions when the contents of the plaintext are identical between entries, even on large entries.