Feature: Rewrote wallet storage to decrease deletion times
The wallet storage took ages, so I rewrote it to use a mix of the password storage method and the existing wallet implementation. There are a bunch of 32-byte blocks that store the encrypted data, a bitmask indicating which blocks are filled, and a 'layout' section indicating which blocks are related to which wallets. These parts allow me to simply tell the bitmask that the block is now free to be overwritten, and have to shift less data down in the layout section.