Feature : Started encrypting comms with AES CTR mode
The responses from the vault for the 'pwd read' and 'wlt read' are now encrypted with AES CTR. The key used is the session key and the nonce and iv are from the nonce exchanged during the initial authentication. This has been added to both the vault and the python client and is working, its not on any of the other commands yet as the sensitive information for other commands travels from the client, so this data will need to be encrypted as well.